test: collapse startup-resync tests into a single package
We no longer share devnets across packages, so there's no need to keep
the post-activation and pre-activation cases in separate sub-packages.
Both tests now live in a single startup_resync_test.go file.
