+++ op-acceptance-tests/tests/supernode/interop/startup_resync/startup_resync_test.go
@@ -14,34 +14,24 @@ import (
 	"github.com/ethereum-optimism/optimism/op-supervisor/supervisor/types"
 )
 
-const (
-	l2BlockTime         = uint64(1)
-	backfillDepth       = 3 * time.Second
-	preRestartFinalized = uint64(5)
-)
-
 // TestSupernodeResyncResumesAtActivation_PostActivation drives a full
 // supernode data-dir wipe after the chain has crossed activation, and
-// asserts that cross-safe keeps advancing post-restart and that the
-// cold-start backfill restored history into the logs DB.
+// asserts cross-safe advances post-restart — proving the sequencer is still
+// producing blocks and the supernode is validating them.
 func TestSupernodeResyncResumesAtActivation_PostActivation(gt *testing.T) {
 	t := devtest.SerialT(gt)
 	sys := presets.NewTwoL2SupernodeInterop(t, 0,
-		presets.WithUniformL2BlockTimes(l2BlockTime),
-		presets.WithInteropLogBackfillDepth(backfillDepth),
+		presets.WithTimeTravelEnabled(),
+		presets.WithInteropLogBackfillDepth(60*time.Second),
 	)
 
 	sys.Supernode.AwaitBackfillCompleted()
 
-	// Setup: let L2 finalized advance several blocks on both chains. On
-	// restart, op-node may drop back as part of its safe start process,
-	// but won't go past the finalized head. With finalized well past
-	// genesis the post-restart cold-start backfill has a real window to
-	// populate, instead of collapsing to empty against a re-recorded
-	// genesis SafeDB entry.
+	// Setup: prove the initial supernode is producing committed cross-safe
+	// entries on both chains.
 	dsl.CheckAll(t,
-		sys.L2ACL.AdvancedFn(types.Finalized, preRestartFinalized, 180),
-		sys.L2BCL.AdvancedFn(types.Finalized, preRestartFinalized, 180),
+		sys.L2ACL.AdvancedFn(types.CrossSafe, 5, 60),
+		sys.L2BCL.AdvancedFn(types.CrossSafe, 5, 60),
 	)
 
 	sys.Supernode.RestartWithFreshDataDir()
@@ -51,24 +41,29 @@ func TestSupernodeResyncResumesAtActivation_PostActivation(gt *testing.T) {
 		sys.L2ACL.AdvancedFn(types.CrossSafe, 1, 60),
 		sys.L2BCL.AdvancedFn(types.CrossSafe, 1, 60),
 	)
-
-	// Verify the cold-start backfill repopulated the logs DB.
-	sys.Supernode.AssertBackfillCovers(backfillDepth, l2BlockTime,
-		sys.L2A.ChainID(), sys.L2B.ChainID())
 }
 
-// TestSupernodeResyncSchedulesAtActivation_PreActivation drives a full
+// preActivationDelay schedules interop activation far enough past genesis
+// that the chain tip stays comfortably pre-activation through the whole
+// pre-activation test, even under CI scheduling variance.
+const preActivationDelay = uint64(300)
+
+// TestSupernodeResyncResumesAtActivation_PreActivation drives a full
 // supernode data-dir wipe while interop is scheduled but not yet active,
-// and asserts that cold-start init parks the verifier at the (future)
-// activation timestamp while cross-safe keeps advancing on both chains.
-func TestSupernodeResyncSchedulesAtActivation_PreActivation(gt *testing.T) {
+// and asserts:
+//
+//   - cold-start init resumes the verifier at the activation timestamp
+//     (which is still in the future), and
+//   - cross-safe keeps advancing on both chains while interop sits idle
+//     waiting for activation.
+//
+// The verifier transition at activation itself is covered by other tests;
+// here we only assert pre-activation cold-start behaviour.
+func TestSupernodeResyncResumesAtActivation_PreActivation(gt *testing.T) {
 	t := devtest.SerialT(gt)
-	// 60-minute delay: ensures the chain never approaches activation during
-	// the test, so we always exercise the genuine pre-activation cold-start
-	// path regardless of CI scheduling variance.
-	sys := presets.NewTwoL2SupernodeInterop(t, 60*60,
-		presets.WithUniformL2BlockTimes(l2BlockTime),
-		presets.WithInteropLogBackfillDepth(backfillDepth),
+	sys := presets.NewTwoL2SupernodeInterop(t, preActivationDelay,
+		presets.WithTimeTravelEnabled(),
+		presets.WithInteropLogBackfillDepth(60*time.Second),
 	)
 
 	sys.Supernode.AwaitBackfillCompleted()
